Allow for work on different deployment models
As noted on the list, ruling discovery completely out of bounds wasn't
the greatest thing ever for DoH.

The intent here is to explicitly keep discussion of implications in
scope, but to avoid having it overwhelm the core work by clearly
indicating that the core protocol elements take priority.

Allowing the group to discuss deployment models and document them should
help.  This extends to allowing the working group to work on mechanisms
in support of those deployment models.  If the group decides that there
is a viable mechanism that depends on some sort of discovery mechanism,
then the definition of the necessary mechanisms can be done under the
charter.

The usual negotiation with Area Directors about the addition of new
milestones remain in full force, so this isn't an unbounded scope.  In
particular, the initial charter won't include explicit milestones for
specific work in this area.