Some amount of rewording
This reworks @ekr's contribution in #5 (the text already said the first
part of his addition).

This also rewords the language on examples a little.  I agree with Chris
that the main thrust of the anti-general purpose thing is server
support.  Turns out, you could use this for general purpose browsing,
including using cookies, if you cared less about linking requests.  That
would hide IP addresses, which is not a negligible thing.  Performance
is worse, certainly, but that is still stymied by the lack of server
support.

Eric O is ALSO right: a two-layer proxy system could use this as an
anonymization system, but the anti-abuse systems would need a lot of
work.

Closes #5.